Upgaded Cheating Chester
Cheating Chester still does not work as he cannot import the main rps file without causing a loop. The only solution i can think of is to import each bot separately for two rounds and If he can win both rounds stick with that bot and assume that he is the bot he is fighting.